ai/sunset/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
torch==2.1.0
transformers==4.35.0
diffusers==0.16.1
scipy==1.11.1
ftfy==6.1.1
accelerate==0.21.0
numpy==1.24.4
pillow==10.0.0
